refactor engine into small parts
